🔨 Wrap content with main tag in subtemplates
All content was previously wrapped in a main tag in the default
template. This moves the main tag into home and errata templates,
allowing for more granular class control. The main tag has the cover
class added to it, while the errata main tag is left as is. The div tag
around the cover also gains the cover class.

Note that the primary button in the home template has also had it's
class changed.